+++ src/main/java/com/saditec/platform/SaditecServiceApplication.java
@@ -6,8 +6,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class SaditecServiceApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(SaditecServiceApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(SaditecServiceApplication.class, args);
+    }
 
 }