+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-#### [0.0.2] - 2024-06-04
+#### [0.0.3] - 2024-06-07
+
+- Added entities for educational and recreational facilities part 1
+- Implemented relationship on entities part 1
+
+#### [0.0.2] - 2024-06-05
 
 - Configure application properties